requirements.txt
discord.py==2.2.3
pip-chill==1.0.3
python-dotenv==1.0.0